Move launch functions into their own file
